% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_for.R
\name{pb_for}
\alias{pb_for}
\title{Use a progress bar with regular for loops}
\usage{
pb_for(format = "[:bar] :percent", width = options("width")[[1]] - 2,
  complete = "=", incomplete = "-", current = ">",
  callback = invisible, clear = TRUE, show_after = 0.2,
  force = FALSE, once = TRUE)
}
\arguments{
\item{format}{The format of the progress bar. A number of tokens can be used
here, see them below. It defaults to "[:bar] :percent", which means that
the progress bar is within brackets on the left, and the percentage is
printed on the right.}

\item{width}{Width of the progress bar.}

\item{complete}{Completion character.}

\item{incomplete}{Incomplete character.}

\item{current}{Current character.}

\item{callback}{Callback function to call when the progress bar finishes.
The progress bar object itself is passed to it as the single parameter.}

\item{clear}{Whether to clear the progress bar on completion.}

\item{show_after}{Amount of time in seconds, after which the progress bar is
shown on the screen. For very short processes, it is probably not worth
showing it at all.}

\item{force}{Whether to force showing the progress bar, even if the given (or default) stream does not seem to support it.}

\item{once}{Wether to return `for` to it's default behavior after it's been
called once.}
}
\description{
A wrapper around the package *progress* to be able to use progress bar with
regular `for` loops conveniently. It forwards all its parameters to
`progress::progress_bar$new()` except for `once`.
}
\examples{
 \dontrun{
  pb_for()
  for (i in 1:10) {
    # DO SOMETHING
    Sys.sleep(0.5)
  }

  pb_for(format = "Working hard: [:bar] :percent :elapsed", callback = ~message("Were'd done!"))
  for (i in 1:10) {
    # DO SOMETHING
    Sys.sleep(0.5)
  }
}
}
